requirements.txt
fastapi
uvicorn
pyinstaller
python-dotenv
